+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # v1.11.8
+# New Features
+- Added a regenerate thumbnails context item to the Scene Folder.
+- Rewire fireplaces should now also remap the tile image.
 # Bug fixes
 - Fix image link for Lightbringer
 # v1.11.7